instaagent/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
python-dotenv==1.0.1
supabase
anthropic
httpx
Pillow==10.3.0
celery==5.3.6
redis==5.0.3
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
razorpay==1.4.1
python-telegram-bot==21.1.1
pytest==8.1.0
pytest-asyncio==0.23.5

# Gmail API — email delivery (OTP + notifications)
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
watchdog==4.0.0
PyYAML==6.0.1
cryptography==42.0.5
email-validator==2.1.0.post1